HADOOP-17461. moving IOStatisticsContext in a different package
